network: moving the Route Filter parser into the parse package
