+++ azurerm/internal/services/network/parse/route_filter.go
@@ -1,4 +1,4 @@
-package network
+package parse
 
 import (
 	"fmt"
