+++ azurerm/internal/services/network/route_filter_resource.go
@@ -12,6 +12,7 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/azure"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/tf"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/clients"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/network/parse"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/tags"
 	azSchema "github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/tf/schema"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/timeouts"
@@ -26,7 +27,7 @@ func resourceArmRouteFilter() *schema.Resource {
 		Delete: resourceArmRouteFilterDelete,
 
 		Importer: azSchema.ValidateResourceIDPriorToImport(func(id string) error {
-			_, err := ParseRouteFilterID(id)
+			_, err := parse.ParseRouteFilterID(id)
 			return err
 		}),
 
@@ -158,7 +159,7 @@ func resourceArmRouteFilterRead(d *schema.ResourceData, meta interface{}) error 
 	ctx, cancel := timeouts.ForRead(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
-	id, err := ParseRouteFilterID(d.Id())
+	id, err := parse.ParseRouteFilterID(d.Id())
 	if err != nil {
 		return err
 	}
@@ -192,7 +193,7 @@ func resourceArmRouteFilterDelete(d *schema.ResourceData, meta interface{}) erro
 	ctx, cancel := timeouts.ForDelete(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
-	id, err := ParseRouteFilterID(d.Id())
+	id, err := parse.ParseRouteFilterID(d.Id())
 	if err != nil {
 		return err
 	}
